requirements.txt
click==8.1.3
Flask==2.3.2
gunicorn==20.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.3.3
-e .
Authlib==0.15.5
flask-login==0.6.2
python-dotenv==1.0.0
spotipy
beautifulsoup4==4.9.1
requests>=2.25
